+++ Sources/Keychain/KeychainStorage.swift
@@ -7,6 +7,25 @@ open class KeychainStorage: DelegatedStorage {
     /// `KeychainStorage` shared instance.
     open class var standard: KeychainStorage { shared }
     private static let shared = KeychainStorage()
+    private var keychainDelegate: KeychainStorageDelegate? { delegate as? KeychainStorageDelegate }
+
+    /// Access group where `StorageData` is in.
+    open var accessGroup: String? {
+        get { keychainDelegate?.accessGroup }
+        set { keychainDelegate?.accessGroup = newValue }
+    }
+
+    /// Whether the `StorageData` can be synchronized.
+    open var synchronizable: Bool {
+        get { keychainDelegate?.synchronizable ?? false }
+        set { keychainDelegate?.synchronizable = newValue }
+    }
+
+    /// When `StorageData` can be accessed in the keychain.
+    open var accessible: CFString {
+        get { keychainDelegate?.accessible ?? kSecAttrAccessibleWhenUnlocked }
+        set { keychainDelegate?.accessible = newValue }
+    }
 
     /**
      Create a `KeychainStorage`.
@@ -28,6 +47,13 @@ open class KeychainStorage: DelegatedStorage {
 
 /// `KeychainStorageDelegate` conforming `StorageDelegate` protocol.
 open class KeychainStorageDelegate: StorageDelegate {
+    /// Access group where `StorageData` is in.
+    open var accessGroup: String?
+    /// Whether the `StorageData` can be synchronized.
+    open var synchronizable = false
+    /// When `StorageData` can be accessed in the keychain.
+    open var accessible: CFString = kSecAttrAccessibleWhenUnlocked
+
     /// Create a `KeychainStorageDelegate`.
     public init() {}
 
@@ -41,7 +67,9 @@ open class KeychainStorageDelegate: StorageDelegate {
      - Returns: `StorageData` for `StoreKey`.
      */
     open func data<D: StorageData>(forKey key: StoreKey) throws -> D? {
-        try read(account: key)
+        try read(account: key,
+                 accessGroup: accessGroup,
+                 synchronizable: synchronizable)
     }
 
     /**
@@ -53,9 +81,13 @@ open class KeychainStorageDelegate: StorageDelegate {
      - Throws: `KeychainError.error`.
      */
     open func set<D: StorageData>(_ data: D?, forKey key: StoreKey) throws {
-        try remove(forKey: key)
+        try? remove(forKey: key)
         if let data = data {
-            try store(data, account: key)
+            try store(data,
+                      account: key,
+                      accessible: accessible,
+                      accessGroup: accessGroup,
+                      synchronizable: synchronizable)
         }
     }
 
@@ -67,7 +99,9 @@ open class KeychainStorageDelegate: StorageDelegate {
      - Throws: `KeychainError.error`.
      */
     open func remove(forKey key: StoreKey) throws {
-        try delete(account: key)
+        try delete(account: key,
+                   accessGroup: accessGroup,
+                   synchronizable: synchronizable)
     }
 
     /**
@@ -76,17 +110,28 @@ open class KeychainStorageDelegate: StorageDelegate {
      - Parameter value: `StorageData` to store.
      - Parameter account: Item's account name.
      - Parameter accessible: When the keychain item is accessible.
+     - Parameter accessGroup: Access group where `StorageData` is in.
+     - Parameter synchronizable: Whether the `StorageData` can be synchronized.
 
      - Throws: `KeychainError.error`.
      */
     open func store<D: StorageData>(_ value: D,
                                     account: String,
-                                    accessible: CFString = kSecAttrAccessibleWhenUnlocked) throws {
-        let query = [kSecClass: kSecClassGenericPassword,
+                                    accessible: CFString,
+                                    accessGroup: String?,
+                                    synchronizable: Bool) throws {
+        var query = [kSecClass: kSecClassGenericPassword,
                      kSecAttrAccount: account,
                      kSecAttrAccessible: accessible,
                      kSecUseDataProtectionKeychain: true,
-                     kSecValueData: value.data] as [String: Any]
+                     kSecValueData: value.data] as [CFString: Any]
+        if let accessGroup = accessGroup {
+            query[kSecAttrAccessGroup] = accessGroup
+        }
+        if synchronizable {
+            query[kSecAttrSynchronizable] = kCFBooleanTrue
+        }
+
         let status = SecItemAdd(query as CFDictionary, nil)
         guard status == errSecSuccess else {
             throw KeychainError.error("Unable to store item: \(status.message)")
@@ -97,16 +142,26 @@ open class KeychainStorageDelegate: StorageDelegate {
      Read `StorageData` for account from the keychain.
 
      - Parameter account: Item's account name.
+     - Parameter accessGroup: Access group where `StorageData` is in.
+     - Parameter synchronizable: Whether the `StorageData` can be synchronized.
 
      - Throws: `KeychainError.error`.
 
      - Returns: `StorageData` for account.
      */
-    open func read<D: StorageData>(account: String) throws -> D? {
-        let query = [kSecClass: kSecClassGenericPassword,
+    open func read<D: StorageData>(account: String,
+                                   accessGroup: String?,
+                                   synchronizable: Bool) throws -> D? {
+        var query = [kSecClass: kSecClassGenericPassword,
                      kSecAttrAccount: account,
                      kSecUseDataProtectionKeychain: true,
-                     kSecReturnData: true] as [String: Any]
+                     kSecReturnData: true] as [CFString: Any]
+        if let accessGroup = accessGroup {
+            query[kSecAttrAccessGroup] = accessGroup
+        }
+        if synchronizable {
+            query[kSecAttrSynchronizable] = kCFBooleanTrue
+        }
 
         var item: CFTypeRef?
         switch SecItemCopyMatching(query as CFDictionary, &item) {
@@ -126,13 +181,24 @@ open class KeychainStorageDelegate: StorageDelegate {
      Delete item for account from the keychain.
 
      - Parameter account: Item's account name.
+     - Parameter accessGroup: Access group where `StorageData` is in.
+     - Parameter synchronizable: Whether the `StorageData` can be synchronized.
 
      - Throws: `KeychainError.error`.
      */
-    open func delete(account: String) throws {
-        let query = [kSecClass: kSecClassGenericPassword,
+    open func delete(account: String,
+                     accessGroup: String?,
+                     synchronizable: Bool) throws {
+        var query = [kSecClass: kSecClassGenericPassword,
                      kSecUseDataProtectionKeychain: true,
-                     kSecAttrAccount: account] as [String: Any]
+                     kSecAttrAccount: account] as [CFString: Any]
+        if let accessGroup = accessGroup {
+            query[kSecAttrAccessGroup] = accessGroup
+        }
+        if synchronizable {
+            query[kSecAttrSynchronizable] = kCFBooleanTrue
+        }
+
         switch SecItemDelete(query as CFDictionary) {
         case errSecItemNotFound, errSecSuccess:
             break